Removed unnecessary interface
Started removing unnecessary interfaces. They don't turn out useful after all.
