+++ src/main/resources/META-INF/gradle-plugins/org.mockito.mockito-release-tools.bintray.properties
@@ -1,1 +1,1 @@
-implementation-class=org.mockito.release.internal.gradle.DefaultBintrayPlugin
+implementation-class=org.mockito.release.gradle.BintrayPlugin
